+++ be/src/main/java/buddyguard/mybuddyguard/jwt/JwtFilter.java
@@ -1,13 +1,13 @@
 package buddyguard.mybuddyguard.jwt;
 
 import buddyguard.mybuddyguard.exception.FilterException;
+import buddyguard.mybuddyguard.jwt.service.TokenService;
 import buddyguard.mybuddyguard.login.dto.CustomOAuth2User;
 import buddyguard.mybuddyguard.login.dto.UserDto;
 import buddyguard.mybuddyguard.login.exception.NotAccessTokenException;
 import buddyguard.mybuddyguard.login.exception.TokenExpiredException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -20,11 +20,11 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 public class JwtFilter extends OncePerRequestFilter {
 
-    private final JwtUtil jwtUtil;
+    private final TokenService tokenService;
     private final AuthenticationEntryPoint authenticationEntryPoint;
 
-    public JwtFilter(JwtUtil jwtUtil, AuthenticationEntryPoint authenticationEntryPoint) {
-        this.jwtUtil = jwtUtil;
+    public JwtFilter(TokenService tokenService, AuthenticationEntryPoint authenticationEntryPoint) {
+        this.tokenService = tokenService;
         this.authenticationEntryPoint = authenticationEntryPoint;
     }
 
@@ -40,12 +40,12 @@ public class JwtFilter extends OncePerRequestFilter {
         }
 
         try {
-            if (jwtUtil.isExpired(accessToken)) {
+            if (tokenService.isExpired(accessToken)) {
                 // 토큰이 만료된 상황
                 throw new TokenExpiredException(HttpStatus.UNAUTHORIZED, "unauthorized token");
             }
 
-            Tokens type = jwtUtil.getTokenType(accessToken);
+            Tokens type = tokenService.getTokenType(accessToken);
             if (!type.equals(Tokens.ACCESS)) {
                 // 엑세스 토큰이 아닌 상황
                 throw new NotAccessTokenException(HttpStatus.UNAUTHORIZED, "unauthorized token");
@@ -56,9 +56,9 @@ public class JwtFilter extends OncePerRequestFilter {
             return;
         }
 
-        Long userId = jwtUtil.getUserId(accessToken);
-        String role = jwtUtil.getRole(accessToken);
-        String name = jwtUtil.getName(accessToken);
+        Long userId = tokenService.getUserId(accessToken);
+        String role = tokenService.getRole(accessToken);
+        String name = tokenService.getName(accessToken);
 
         UserDto userDto = new UserDto();
         userDto.setId(userId);
