--- be/src/main/java/buddyguard/mybuddyguard/jwt/JwtUtil.java
@@ -1,78 +1,0 @@
-package buddyguard.mybuddyguard.jwt;
-
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.Jwts.SIG;
-import java.nio.charset.StandardCharsets;
-import java.time.Instant;
-import java.util.Date;
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-/**
- * JWT 토큰과 관련된 작업들을 진행하는 클래스
- */
-@Component
-public class JwtUtil {
-    private final SecretKey secretKey;
-
-    public JwtUtil(@Value("${spring.jwt.secret}") String secret) {
-        this.secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8),
-                SIG.HS256.key().build().getAlgorithm());
-    }
-
-    public Long getUserId(String token) {
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload()
-                .get("userId", Long.class);
-    }
-
-    public String getRole(String token) {
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload()
-                .get("role", String.class);
-    }
-
-    public String getName(String token) {
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload()
-                .get("name", String.class);
-    }
-
-    /**
-     * 토큰의 유효기간이 만료되었는지 확인합니다
-     * @param token
-     * @return
-     */
-    public Boolean isExpired(String token) {
-        try {
-            Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token);
-            return false;
-        } catch (ExpiredJwtException e) {
-            return true;
-        }
-    }
-
-    public Tokens getTokenType(String token) {
-        String stringTokens = Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload()
-                .get("tokenType", String.class);
-        return Tokens.valueOf(stringTokens);
-    }
-
-    /**
-     * JWT 토큰을 생성합니다
-     * @param userId
-     * @param role
-     * @param expiredSeconds
-     * @return
-     */
-    public String createJwt(Long userId, String role, Tokens tokenType, Long expiredSeconds) {
-        return Jwts.builder()
-                .claim("userId", userId)
-                .claim("role", role)
-                .claim("tokenType", tokenType)
-                .issuedAt(Date.from(Instant.now()))
-                .expiration(Date.from(Instant.now().plusSeconds(expiredSeconds)))
-                .signWith(secretKey)
-                .compact();
-    }
-}
